refactor(site): isolate sitemap generation
Move sitemap creation and its regression test into a dedicated module. Keep Vitest source discovery explicit without loading unrelated browser or external-data modules.
