+++ vite.config.ts
@@ -124,7 +124,7 @@ export default defineConfig({
 					includeSource: [
 						'src/lib/**/*.ts',
 						'src/site/{assets,content-assets,dev-routes,dev-server}.ts',
-						'src/site/{generate,page-styles,pages}.ts',
+						'src/site/{generate,page-styles,pages,sitemap}.ts',
 						'packages/content/src/{artifact,blog,island-renderer,islands,ogp-snapshots,paths,tweet-snapshots}.ts',
 						'packages/content/src/blog/**/*.ts',
 						'packages/content/src/markdown/**/*.ts',
